+++ wo/cli/plugins/site_functions.py
@@ -296,7 +296,7 @@ def setupwordpress(self, data):
         Log.debug(self, "Generating wp-config for WordPress Single site")
         Log.debug(self, "bash -c \"php {0} --allow-root "
                   .format(WOVariables.wo_wpcli_path) +
-                  "core config " +
+                  "config create " +
                   "--dbname=\'{0}\' --dbprefix=\'{1}\' --dbuser=\'{2}\' "
                   "--dbhost=\'{3}\' "
                   .format(data['wo_db_name'], wo_wp_prefix,
@@ -308,7 +308,7 @@ def setupwordpress(self, data):
         try:
             if WOShellExec.cmd_exec(self, "bash -c \"php {0} --allow-root"
                                     .format(WOVariables.wo_wpcli_path) +
-                                    " core config " +
+                                    " config create " +
                                     "--dbname=\'{0}\' --dbprefix=\'{1}\' "
                                     "--dbuser=\'{2}\' --dbhost=\'{3}\' "
                                     .format(data['wo_db_name'], wo_wp_prefix,
@@ -317,12 +317,9 @@ def setupwordpress(self, data):
                                             ) +
                                     "--dbpass=\'{0}\' "
                                     "--extra-php<<PHP \n"
-                                    "{1} {redissalt}\nPHP\""
+                                    "\n{1}\nPHP\""
                                     .format(data['wo_db_pass'],
-                                            "\n\ndefine(\'WP_DEBUG\', false);",
-                                            redissalt="\n\ndefine( \'WP_CACHE_KEY_SALT\', \'{0}:\' );"
-                                                      .format(wo_domain_name) if data['wpredis']
-                                                      else ''),
+                                            "\ndefine(\'WP_DEBUG\', false);"),
                                     log=False
                                     ):
                 pass
@@ -330,11 +327,17 @@ def setupwordpress(self, data):
                 raise SiteError("generate wp-config failed for wp single site")
         except CommandExecutionError as e:
             raise SiteError("generate wp-config failed for wp single site")
+
+        WOShellExec.cmd_exec(self, "bash -c \"php {0} --allow-root "
+                             .format(WOVariables.wo_wpcli_path) +
+                             "config set WP_CACHE_KEY_SALT "
+                             "\'{0}:\'\"".format(wo_domain_name))
+
     else:
         Log.debug(self, "Generating wp-config for WordPress multisite")
         Log.debug(self, "bash -c \"php {0} --allow-root "
                   .format(WOVariables.wo_wpcli_path) +
-                  "core config " +
+                  "config create " +
                   "--dbname=\'{0}\' --dbprefix=\'{1}\' --dbhost=\'{2}\' "
                   .format(data['wo_db_name'], wo_wp_prefix, data['wo_db_host']) +
                   "--dbuser=\'{0}\' --dbpass=\'{1}\' "
@@ -348,21 +351,19 @@ def setupwordpress(self, data):
         try:
             if WOShellExec.cmd_exec(self, "bash -c \"php {0} --allow-root"
                                     .format(WOVariables.wo_wpcli_path) +
-                                    " core config " +
+                                    " config create " +
                                     "--dbname=\'{0}\' --dbprefix=\'{1}\' "
                                     "--dbhost=\'{2}\' "
                                     .format(data['wo_db_name'], wo_wp_prefix,
                                             data['wo_db_host']) +
                                     "--dbuser=\'{0}\' --dbpass=\'{1}\' "
-                                    "--extra-php<<PHP \n {2} {3} {redissalt}\nPHP\""
+                                    "--extra-php<<PHP \n "
+                                    "\n{2} {3}\nPHP\""
                                     .format(data['wo_db_user'],
                                             data['wo_db_pass'],
                                             "\ndefine(\'WPMU_ACCEL_REDIRECT\',"
                                             " true);",
-                                            "\n\ndefine(\'WP_DEBUG\', false);",
-                                            redissalt="\n\ndefine( \'WP_CACHE_KEY_SALT\', \'{0}:\' );"
-                                            .format(wo_domain_name) if data['wpredis']
-                                            else ''),
+                                            "\ndefine(\'WP_DEBUG\', false);"),
                                     log=False
                                     ):
                 pass
@@ -370,6 +371,11 @@ def setupwordpress(self, data):
                 raise SiteError("generate wp-config failed for wp multi site")
         except CommandExecutionError as e:
             raise SiteError("generate wp-config failed for wp multi site")
+
+        WOShellExec.cmd_exec(self, "bash -c \"php {0} --allow-root "
+                             .format(WOVariables.wo_wpcli_path) +
+                             "config set WP_CACHE_KEY_SALT "
+                             "\'{0}:\'\"".format(wo_domain_name))
 
     # WOFileUtils.mvfile(self, os.getcwd()+'/wp-config.php',
     #                   os.path.abspath(os.path.join(os.getcwd(), os.pardir)))
@@ -1279,7 +1285,7 @@ def setupLetsEncrypt(self, wo_domain_name):
                                        "'/etc/letsencrypt/config' "
                                        "--issue "
                                        "-d {0} -d www.{0} -w /var/www/html "
-                                       "-k ec-384 --force"
+                                       "-k ec-384 -f"
                                        .format(wo_domain_name))
     else:
         Log.info(self, "Issuing SSL cert with acme.sh")
@@ -1288,7 +1294,7 @@ def setupLetsEncrypt(self, wo_domain_name):
                                    "'/etc/letsencrypt/config' "
                                    "--issue "
                                    "-d {0} -d www.{0} -w /var/www/html "
-                                   "-k ec-384"
+                                   "-k ec-384 -f"
                                    .format(wo_domain_name))
 
     if ssl:
@@ -1308,7 +1314,8 @@ def setupLetsEncrypt(self, wo_domain_name):
                                             "{0}/{1}/fullchain.pem "
                                             "--ca-file {0}/{1}/ca.pem "
                                             "--reloadcmd "
-                                            "\"service nginx restart\" "
+                                            "\"nginx -t && "
+                                            "service nginx restart\" "
                                             .format(WOVariables.wo_ssl_live,
                                                     wo_domain_name))
             Log.info(
@@ -1363,7 +1370,7 @@ def setupLetsEncryptSubdomain(self, wo_domain_name):
                                        "'/etc/letsencrypt/config' "
                                        "--issue "
                                        "-d {0} -w /var/www/html "
-                                       "-k ec-384 --force"
+                                       "-k ec-384 -f"
                                        .format(wo_domain_name))
     else:
         Log.info(self, "Issuing SSL cert with acme.sh")
@@ -1372,7 +1379,7 @@ def setupLetsEncryptSubdomain(self, wo_domain_name):
                                    "'/etc/letsencrypt/config' "
                                    "--issue "
                                    "-d {0} -w /var/www/html "
-                                   "-k ec-384"
+                                   "-k ec-384 -f"
                                    .format(wo_domain_name))
     if ssl:
 
@@ -1391,7 +1398,7 @@ def setupLetsEncryptSubdomain(self, wo_domain_name):
                                             "{0}/{1}/fullchain.pem "
                                             "--ca-file {0}/{1}/ca.pem "
                                             "--reloadcmd "
-                                            "\"service nginx restart\" "
+                                            "\"nginx -t && service nginx restart\" "
                                             .format(WOVariables.wo_ssl_live,
                                                     wo_domain_name))