+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### v3.9.6 - [Unreleased]
 
+#### Fixed
+
+- WP_CACHE_KEY_SALT set twice
+
 ### v3.9.5.1 - 2019-05-10
 
 #### Fixed